🤖 Refactor(#56): move get-kakao-token api function to user.ts
